Player Attack

+++ Assets/scripts/Player/IDamageable.cs
@@ -5,5 +5,5 @@ using UnityEngine;
 public interface IDamageable
 {
 
-    void TakeDamage(int damageAmount);
+    void TakeDamage(float damageAmount);
 }
